move, add rule
