+++ apps/dotcom/client/src/tla/components/TlaFileMenu/TlaFileMenu.tsx
@@ -17,11 +17,11 @@ import {
 	useToasts,
 } from 'tldraw'
 import { TldrawApp } from '../../app/TldrawApp'
-import { defineMessages, useIntl } from '../../app/i18n'
 import { useApp } from '../../hooks/useAppState'
 import { useIsFileOwner } from '../../hooks/useIsFileOwner'
 import { TLAppUiEventSource, useTldrawAppUiEvents } from '../../utils/app-ui-events'
 import { copyTextToClipboard } from '../../utils/copy'
+import { defineMessages, useIntl } from '../../utils/i18n'
 import { getFilePath, getShareableFileUrl } from '../../utils/urls'
 import { TlaDeleteFileDialog } from '../dialogs/TlaDeleteFileDialog'
