+++ apps/dotcom/client/src/tla/components/TlaFileShareMenu/Tabs/TlaInviteTab.tsx
@@ -1,12 +1,12 @@
 import { TlaFile } from '@tldraw/dotcom-shared'
 import { useCallback } from 'react'
 import { useEditor, useValue } from 'tldraw'
-import { F, defineMessages, useIntl } from '../../../app/i18n'
 import { useApp } from '../../../hooks/useAppState'
 import { useIsFileOwner } from '../../../hooks/useIsFileOwner'
 import { useTldrawUser } from '../../../hooks/useUser'
 import { useTldrawAppUiEvents } from '../../../utils/app-ui-events'
 import { copyTextToClipboard } from '../../../utils/copy'
+import { F, defineMessages, useIntl } from '../../../utils/i18n'
 import { getShareableFileUrl } from '../../../utils/urls'
 import { TlaSelect } from '../../TlaSelect/TlaSelect'
 import { TlaSwitch } from '../../TlaSwitch/TlaSwitch'
