+++ apps/dotcom/client/src/tla/components/TlaFileShareMenu/Tabs/TlaPublishTab.tsx
@@ -3,10 +3,10 @@ import { TlaFile } from '@tldraw/dotcom-shared'
 import { useCallback, useState } from 'react'
 import { useParams } from 'react-router-dom'
 import { useEditor } from 'tldraw'
-import { F, FormattedRelativeTime } from '../../../app/i18n'
 import { useApp } from '../../../hooks/useAppState'
 import { useTldrawAppUiEvents } from '../../../utils/app-ui-events'
 import { copyTextToClipboard } from '../../../utils/copy'
+import { F, FormattedRelativeTime } from '../../../utils/i18n'
 import { getShareablePublishUrl } from '../../../utils/urls'
 import { TlaButton } from '../../TlaButton/TlaButton'
 import { TlaSwitch } from '../../TlaSwitch/TlaSwitch'
