+++ apps/dotcom/client/src/tla/components/TlaSidebar/TlaSidebar.tsx
@@ -2,12 +2,12 @@ import classNames from 'classnames'
 import { ReactElement, memo, useCallback, useEffect, useRef, useState } from 'react'
 import { Link, useNavigate, useParams } from 'react-router-dom'
 import { TldrawUiInput, preventDefault, useValue } from 'tldraw'
-import { F, defineMessages, useIntl } from '../../app/i18n'
 import { useApp } from '../../hooks/useAppState'
 import { useIsFileOwner } from '../../hooks/useIsFileOwner'
 import { useTldrFileDrop } from '../../hooks/useTldrFileDrop'
 import { TLAppUiEventSource, useTldrawAppUiEvents } from '../../utils/app-ui-events'
 import { getRelevantDates } from '../../utils/dates'
+import { F, defineMessages, useIntl } from '../../utils/i18n'
 import { getLocalSessionState, updateLocalSessionState } from '../../utils/local-session-state'
 import { getFilePath } from '../../utils/urls'
 import { TlaAccountMenu } from '../TlaAccountMenu/TlaAccountMenu'
