+++ apps/dotcom/client/src/tla/layouts/TlaAnonLayout/TlaAnonLayout.tsx
@@ -2,13 +2,13 @@ import { SignedOut } from '@clerk/clerk-react'
 import classNames from 'classnames'
 import { ReactNode } from 'react'
 import { Link } from 'react-router-dom'
-import { defineMessages, F, useIntl } from '../../app/i18n'
 import { TlaAccountMenu } from '../../components/TlaAccountMenu/TlaAccountMenu'
 import { TlaIcon } from '../../components/TlaIcon/TlaIcon'
 import { TlaSignInButton } from '../../components/TlaSignInButton/TlaSignInButton'
 import { TlaSignUpButton } from '../../components/TlaSignUpButton/TlaSignUpButton'
 import { usePreventAccidentalDrops } from '../../hooks/usePreventAccidentalDrops'
 import { useTldrawAppUiEvents } from '../../utils/app-ui-events'
+import { defineMessages, F, useIntl } from '../../utils/i18n'
 import styles from './anon.module.css'
 
 const messages = defineMessages({
