+++ apps/dotcom/client/src/tla/utils/i18n.tsx
@@ -1,3 +1,5 @@
+/* eslint-disable no-restricted-imports */
+
 import {
 	FormattedMessage,
 	IntlConfig,
